Restructure lib and throttle topics
